+++ packages/ghana-bdr/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-ghana-bdr
 
+## 0.1.11 - 14 July 2025
+
+### Patch Changes
+
+- Updated dependencies \[9b5a4f8]
+  - @openfn/language-common@3.0.2
+
 ## 0.1.10 - 10 July 2025
 
 ### Patch Changes