+++ packages/senaite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-senaite
 
+## 1.0.6 - 14 July 2025
+
+### Patch Changes
+
+- Updated dependencies \[9b5a4f8]
+  - @openfn/language-common@3.0.2
+
 ## 1.0.5 - 10 July 2025
 
 ### Patch Changes